+++ src/main/java/csx55/hadoop/QuestionFour/MapperOne.java
@@ -3,15 +3,16 @@ package csx55.hadoop.QuestionFour;
 import java.io.IOException;
 
 import org.apache.commons.beanutils.converters.IntegerArrayConverter;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
 //uses both data sets
 //value = end_of_fad_in + (duration - start_of_fade_out)
-public class MapperOne extends Mapper<Object, Text, Text, IntWritable> {
+public class MapperOne extends Mapper<Object, Text, Text, FloatWritable> {
 
-    private IntWritable fadingTime = new IntWritable();
+    private FloatWritable fadingTime = new FloatWritable();
     private Text artistID = new Text();
 
     @Override
@@ -21,7 +22,7 @@ public class MapperOne extends Mapper<Object, Text, Text, IntWritable> {
         //value = fading time
         
         String[] attributes = value.toString().split("\\|");
-        int calculatedFadingTime = Integer.parseInt(attributes[5]) + (Integer.parseInt(attributes[4]) - Integer.parseInt(attributes[12]));
+        float calculatedFadingTime = Float.parseFloat(attributes[5]) + (Float.parseFloat(attributes[4]) - Float.parseFloat(attributes[12]));
 
         fadingTime.set(calculatedFadingTime); //end_of_fad_in + (duration - start_of_fade_out)
         artistID.set(attributes[0]); //artistID
